% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots.R
\name{add_boundary_knots}
\alias{add_boundary_knots}
\title{Add boundary knots to interior knots.}
\usage{
add_boundary_knots(x, interior_knots, degree = 3, tiny = 1e-05)
}
\description{
Add boundary knots to interior knots.
}
