% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_basis.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{Create design matrix of B-spline basis}
\usage{
create_design_matrix(x_values, knots, degree)
}
\description{
Create design matrix of B-spline basis
}
