% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pspline.R
\name{fit_pspline}
\alias{fit_pspline}
\title{Fitting P-spline model}
\usage{
fit_pspline(
  x_values,
  y_values,
  interior_knots,
  degree = 3,
  lambda = 0.01,
  diff_order = 2
)
}
\arguments{
\item{x_values}{Numeric vector of x (input)}

\item{y_values}{Numeric vector of y (response)}

\item{interior_knots}{Vector of interior knots}

\item{degree}{Degree of the B-spline basis (default = 3)}

\item{lambda}{Smoothing parameter (default = 1e-2)}

\item{diff_order}{Order of the difference penalty (default = 2)}
}
\value{
List contains the spline coefficients, knots, degree, lambda, penalty matrix.
}
\description{
This function computes the coefficients P-spline estimator.
}
\examples{
set.seed(924)
n= 100
x_values= sort(runif(n, 0, 1))
y_values= sin(2*pi*x_values) + cos(4*pi*x_values) + rnorm(n, sd= 0.2)
# spline degree specification
degree= 3
# knot generation
num_interior= 5
interior_knots= knots_quantile(x_values, num_interior)
# setting the value of lambda and the difference penalty order
lambda= 1e-2
diff_order= 2
# model fitting
model= fit_pspline(x_values, y_values, interior_knots, degree, lambda, diff_order)
print(model)


}
