% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots.R
\name{knots_quantile}
\alias{knots_quantile}
\title{The function of interior knots sequence}
\usage{
knots_quantile(x, dimension, degree = 3)
}
\arguments{
\item{x}{Numeric vector of the data}

\item{dimension}{Number of the basis functions}

\item{degree}{Degree of spline (default= 3)}
}
\value{
Numeric vector with interior knots sequence
}
\description{
This function computes a knot sequence using the quantiles of the data.
}
\examples{
set.seed(924)
x = runif(100, 0, 1)
knots = knots_quantile(x, 5, 3)
print(knots)
}
