% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pspline.R
\name{plot_pspline}
\alias{plot_pspline}
\title{Plotting P-spline with scatter plots}
\usage{
plot_pspline(x_values, y_values, model, grid_x)
}
\arguments{
\item{x_values, y_values}{Numeric vector}

\item{model}{Function that fit P-spline with given data points}

\item{grid_x}{Numeric vector with a grid of evaluation points}
}
\description{
This function provides a plot of the P-spline.
}
