% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pspline.R
\name{predict_pspline}
\alias{predict_pspline}
\title{Predicting the values of y for given x based on P-spline}
\usage{
predict_pspline(model, new_x)
}
\arguments{
\item{model}{Function that fit P-spline with given data points}

\item{new_x}{Numeric vector expressing a grid of evaluation points}
}
\value{
Numeric vector with predicted values at new_x
}
\description{
This function computes the predicted values of y for given x based on P-spline.
}
